+++ Development/Source/model_entity_setters.cpp
@@ -16,10 +16,10 @@ void ModelEntity::addPart(const string& value)
 
 void ModelEntity::updateModelMatrix()
 {
-	for (unsigned int i = 0; i < _modelMatrices.size(); i++)
+	for (size_t i = 0; i < _modelMatrices.size(); i++)
 	{
 		// Instanced modelpart cannot have a modelmatrix position
-		if (!getRenderBuffers().empty() && getRenderBuffer(i)->isInstanced())
+		if (!getRenderBuffers().empty() && getRenderBuffer(static_cast<unsigned int>(i))->isInstanced())
 		{
 			_translations[i] = Vec3(0.0f);
 		}
